Add new private subnet and hosts file configuration.
The configuration line "private [subnet <list>] [hosts <file>]" can be
used to optionally provide a list of private subnets and a hosts file
for local use with the NCP.

The default is to consider 376 as a private, local subnet.  The
"subnet" keyword can change this using a list of 8-bit octal number
separated by commas.  The list can have no whitespace.

The NCP will usually resolve host names using DNS.  the "hosts"
keyword can specify a file which contains local hostname-address
mappings.  The format of the file is similar to the standard
/etc/hosts file:

A line that begins with "#" is ignored.  Lines start with an address,
followed by whitespace and a list of hostnames.  An address is an
octal Chaosnet node number.  The list of hostnames is a sequence of
strings separated by whitespace.