Firewall functionality (#44)
A simple firewall is implemented in cbridge, where you can define how to handle connection requests (RFC and BRD packets).

It can be used e.g. to restrict access to services running on your cbridge using the NCP interface, or if your cbridge is in a "gateway position", to filter packets passing through it (i.e. not necessarily with the cbridge as final destination).